Fix Bug #76 Add support to set the password for Root user
Currently the root user is at the Wildfly container
is unable to set up the password.

Deleting the previous root password from the docker image creation
after ran all stuff.

With this change, the root user could be user without require password
through ` su ` command and the change the password.

Signed-off-by: Felipe de Jesus Ruiz Garcia <me@tranzemc.com>