Update group setting
Set the jboss user's group to 0 so that when the container runs in OpenShift, its user
(which would not be jboss) would still belong to the same root group.

This also simplifies the Dockerfile instructions as the installation of WildFly
can be done directly by the jboss user (instead of root) and there is no need
to change the whole installation ownership to jboss:0 as it will be done automatically.

Signed-off-by: Jeff Mesnil <jmesnil@redhat.com>